+++ src/main/java/datastructures/interfaces/LinkedList.java
@@ -2,9 +2,9 @@ package datastructures.interfaces;
 
 public interface LinkedList<E> {
     void add(E element);
-    void add(long index, E element);
-    E get(long index);
-    void remove(long index);
-    long size();
+    void add(int index, E element);
+    E get(int index);
+    void remove(int index);
+    int size();
     String toString();
 }
